Prefer source so dependency tests are installed
Composer installs dependencies from a distribution file by default,
which is usually generated via the likes of `git export` - this means
files listed as export-ignore in that repository's .gitattributes
is ignored, which often includes unit tests are they are not necessary
for production runtime. However during a travis CI test we want to run
these tests as this is a recipe we're testing, not a module project.
Thus we need to include the tests - luckily composer accomodates this
by using the prefer-source flag when updating.