FIX Don't apply widgets to all pages by default
This functionality will be both surprising and for the most part unwanted. Widgets were and are originally intended to complement blog's functionality by rendering things in a side bar such as tag clouds and archive calendars - so it still makes sense to apply the extension to blogs & blog posts by default.