FIX enable CI tests to pass by including theme
Some of the tests (particularly Blog module) rely on the presence of a
theme (or in the least a Page template that supports $Layout) in order to
pass (by asserting that specified content is in the body). As recipe-blog
does not include a theme, this was seeing the tests fail - so fix this by
adjusting the TravisCI configuration file accordingly.